net: move RelayTransaction and the relay-message cache into net_processing
Fourth step of the CConnman backport series (issue #2558, PR 2b). After PR 2a
moved ProcessMessage/SendMessages into net_processing, the relay-message cache
(mapRelay / cs_mapRelay / vRelayExpiration) had its two accessors split across
net.cpp (RelayTransaction) and net_processing.cpp (the SendMessages getdata
loop). Consolidate both there:

- Move RelayTransaction (both overloads) from net.cpp to net_processing.cpp,
  byte-identical; declarations move from net.h to net_processing.h.
- Move mapRelay / cs_mapRelay / vRelayExpiration from net.{h,cpp} into
  net_processing.cpp. Both accessors now live in this TU, so they become
  file-static and the net.h externs are removed.
- RelayTransaction's external callers (rpc/htlc.cpp, rpc/rawtransaction.cpp,
  wallet/wallet.cpp) include net_processing.h; they previously saw the
  declaration transitively through main.h -> net.h.

No behavioral change. RelayInventory stays inline in net.h and is reached from
net_processing.cpp unchanged. mapAlreadyAskedFor is intentionally left in
net.{h,cpp} (CNode::AskFor uses it inline; moving it would invert the
net.h -> net_processing layering) and migrates later with PeerManager.
