+++ src/rpc/htlc.cpp
@@ -6,6 +6,7 @@
 #include "init.h"
 #include <key_io.h>
 #include "main.h"
+#include "net_processing.h"
 #include "policy/fees.h"
 #include "primitives/transaction.h"
 #include "rpc/protocol.h"
